Add request and response types for APIs and methods required by `terraform plan` command
A cusrosy inspection of the [terraform](https://github.com/hashicorp/terraform) code base reveals that the following APIs and methods are used when running `terraform plan` with a `remote` backend:

- [Configuration Versions API](https://developer.hashicorp.com/terraform/cloud-docs/api-docs/configuration-versions)
  - [Show a Configuration Version](https://developer.hashicorp.com/terraform/cloud-docs/api-docs/configuration-versions#show-a-configuration-version)
  - [Create a Configuration Version](https://developer.hashicorp.com/terraform/cloud-docs/api-docs/configuration-versions#create-a-configuration-version)
  - [Upload Configuration Files](https://developer.hashicorp.com/terraform/cloud-docs/api-docs/configuration-versions#upload-configuration-files)
- [Plans API](https://developer.hashicorp.com/terraform/cloud-docs/api-docs/plans)
  - [Show a plan](https://developer.hashicorp.com/terraform/cloud-docs/api-docs/plans#show-a-plan)
- [Runs API](https://developer.hashicorp.com/terraform/cloud-docs/api-docs/run)
  - [Get run details](https://developer.hashicorp.com/terraform/cloud-docs/api-docs/run#get-run-details)
  - [Create a Run](https://developer.hashicorp.com/terraform/cloud-docs/api-docs/run#create-a-run)

This commit adds request and response types for these APIs and methods, as well as interfaces and default implementations that return "not implemented" errors.

Unit tests have been written to ensure that requests and responses are correctly unmarshaled and marshaled, respectively.

Certain timestamps use a format that is similar to Go's [RFC3339](https://cs.opensource.google/go/go/+/refs/tags/go1.21.4:src/time/format.go;l=111), but replace the `Z` with a `+` or `-` sign. A new type `RFC3339Plus` has been created with a custom `MarshalJSON` method to ensure that these timestamps are marshaled correctly. A custom unmarshaler is not necessary as these timestamps are already correctly unmarshaled.

Also added `Makefile` with `test` directive and removed unnecessary packages from `go.mod` and `go.sum`.