Implement minimal API endpoints to support `terraform init`
You can now run `terraform init` on a workspace with the following configuration:

```hcl
terraform {
  backend "remote" {
    hostname = "localhost:8443"
    organization = "default"

    token = "token"

    workspaces {
      name = "default"
    }
  }
}
```

A Terraform workspace with the above configuration now exists at `examples/terraform` and can be used to test functionality.

Must use `"default"` for the `organization` and `workspaces` name. Any token can be used.

Terraform CLI expects the `hostname` to support SSL. We have added SSL support to the API with the following flags:

- `-port-ssl`: port to listen on for SSL
- `-ssl-cert-file`: SSL certificate file
- `-ssl-key-file`: SSL key file

You can run `make certificate` to generate a certificate for `localhost` and add it to the trusted certificates on your machine. Requires [mkcert](https://github.com/FiloSottile/mkcert) to be installed. The `cert.pem` and `key.pem` files are placed in a top-level `cert` folder, and this is where the API looks for these files by default.

Unit tests have been written for marshaling and unmarshaling of new response and request types, respectively.

A `.gitignore` file has been added with the following contents:

- [Terraform.gitignore](https://github.com/github/gitignore/blob/main/Terraform.gitignore)
- `cert/` for certificate generated by `make certificate`