Implement minimal API endpoints to support `terraform plan`
You can now run `terraform plan` using the Terraform workspaces in the `examples/terraform` folder.

The plan will error every time as we are not actually running a plan. However, the server will appropriately respond to the `terraform` client, which gracefully produces an error.

We use [sqlc](https://sqlc.dev/) to manage database schemas and compile SQL queries into type-safe code. Use `make migrate-up` and `make migrate-down` to migrate the database schema.

Right now we are only creating three tables:

- `tfe_configuration_version`
- `tfe_run`
- `tfe_plan`

Prepending `tfe_` to a table name indicates that it stores data for a Terraform Enterprise API schema. Other tables may use different names and prefixes to indicate that they are specifically for OpenAtlas functionality.

We have added [Asynq](https://github.com/hibiken/asynq) as a dependency to handle asynchronous tasks. Right now it is only responsible for processing a run and changing the run and plan to an `errored` state.

We have added a `Dockerfile` to compile OpenAtlas, as well as a `docker-compose.yml` that runs the following containers:

- `api`: OpenAtlas
- `db`: PostgreSQL
- `redis`: Redis (for Asynq)
- `asynqmon`: [asynqmon](https://github.com/hibiken/asynqmon) (to monitor Asynq)

Use `make up` and `make down` to start and stop the `docker-compose.yml` containers.

Future improvements:

- Simplify `internal/api/id.go` to reduce boilerplate
- Move asynchronouse task processing to separate command (so the webserver and workers can run independently)
- Reduce boilerplate for API endpoints
- Gracefully handle termination
- Fix unit tests (not compiling at this time)
- Actually run a Terraform plan